Modified to be less resource intensive
After running many benchmarks, the previous version was found to be
faster, at least up to the tested 1000 kanji, regardless of the
parameters this new version was run with. However, I believe that the
larger the data set, the better this version will be as the previous
version slowed itself down by having too many concurrent threads open.

If one's computer can handle it, the previous version will likely be
faster, but, again, having too many threads for the computer to handle
will actually slow computation.